Refactor code in test_template.py, __main__.py, and utils.py to improve code readability and maintainability
