+++ src/backend/base/langflow/__main__.py
@@ -9,6 +9,12 @@ import click
 import httpx
 import typer
 from dotenv import load_dotenv
+from langflow.main import setup_app
+from langflow.services.database.utils import session_getter
+from langflow.services.deps import get_db_service
+from langflow.services.utils import initialize_services
+from langflow.utils.logger import configure, logger
+from langflow.utils.util import update_settings
 from multiprocess import Process, cpu_count  # type: ignore
 from packaging import version as pkg_version
 from rich import box
@@ -16,12 +22,6 @@ from rich import print as rprint
 from rich.console import Console
 from rich.panel import Panel
 from rich.table import Table
-
-from langflow.main import setup_app
-from langflow.services.database.utils import session_getter
-from langflow.services.deps import get_db_service, get_settings_service
-from langflow.services.utils import initialize_services, initialize_settings_service
-from langflow.utils.logger import configure, logger
 
 console = Console()
 
@@ -66,36 +66,6 @@ def set_var_for_macos_issue():
         # https://stackoverflow.com/questions/75747888/uwsgi-segmentation-fault-with-flask-python-app-behind-nginx-after-running-for-2 # noqa
         os.environ["no_proxy"] = "*"  # to avoid error with gunicorn
         logger.debug("Set OBJC_DISABLE_INITIALIZE_FORK_SAFETY to YES to avoid error")
-
-
-def update_settings(
-    config: str,
-    cache: Optional[str] = None,
-    dev: bool = False,
-    remove_api_keys: bool = False,
-    components_path: Optional[Path] = None,
-    store: bool = True,
-):
-    """Update the settings from a config file."""
-
-    # Check for database_url in the environment variables
-    initialize_settings_service()
-    settings_service = get_settings_service()
-    if config:
-        logger.debug(f"Loading settings from {config}")
-        settings_service.settings.update_from_yaml(config, dev=dev)
-    if remove_api_keys:
-        logger.debug(f"Setting remove_api_keys to {remove_api_keys}")
-        settings_service.settings.update_settings(REMOVE_API_KEYS=remove_api_keys)
-    if cache:
-        logger.debug(f"Setting cache to {cache}")
-        settings_service.settings.update_settings(CACHE=cache)
-    if components_path:
-        logger.debug(f"Adding component path {components_path}")
-        settings_service.settings.update_settings(COMPONENTS_PATH=components_path)
-    if not store:
-        logger.debug("Setting store to False")
-        settings_service.settings.update_settings(STORE=False)
 
 
 @app.command()
